requirements.txt
streamlit==1.25.0
python-dotenv==1.2.1
openai==1.62.0
pydantic==1.10.12
langchain==0.2.14
langchain-core==0.2.33
langchain-community==0.2.12
langchain-openai==0.1.22
langchain-nvidia-ai-endpoints==0.2.2
langchain-text-splitters==0.2.2
faiss-cpu==1.13.2
pypdf==6.6.0
